pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["None"]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^1.4.0"
numpy = "^1.22.2"
beautifulsoup4 = "^4.10.0"
lxml = "^4.7.1"
PySocks = "^1.7.1"
stem = "^1.8.0"
click = "^8.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
